Save OK ENSEMBL gene IDs to a file.
When ENSEMBL Gene IDs are "OK", they get written to a file. This file
can be used on subsequent runs of the program as an alternative to
querying ENSEMBL, which is slow. This should speed up development on
this project, since using the file will prevent a large number of
queries from being made.